+++ internal/observability/testdata/prom.txt
